further tweaking of mobile login flow
