+++ app/lib/Events/UniLinkEvent.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:threebotlogin/services/socketService.dart';
-
 class UniLinkEvent {
   Uri link;
   BuildContext context;
-  BackendConnection connection;
 
-  UniLinkEvent(this.link, this.context, this.connection);
+  UniLinkEvent(this.link, this.context);
 }
